Simplified bulk downloads and search query retries when unsuccessful.
This is mostly compatible with original behavior, but it changes input arguments to be multiple queries instead of one. If you want to do a query with spaces, use quotes around the argument. The query is also not treated as exact search anymore to allow for more chance of finding album in search.